Antialiasing and supersampling working.

+++ ft_utils.c
@@ -6,7 +6,7 @@
 /*   By: rserban <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/01/17 12:28:41 by rserban           #+#    #+#             */
-/*   Updated: 2015/03/09 18:24:13 by rserban          ###   ########.fr       */
+/*   Updated: 2015/03/10 16:14:14 by rserban          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	mem_error(void)
 	exit(-1);
 }
 
-void	get_sx_sy(float *sx, float *sy, float x, float y)
+void	get_sx_sy(float *sx, float *sy, int x, int y)
 {
 	if (WIN_WIDTH > WIN_HEIGHT)
 	{
